chore: merge match & transform into worker
